完成首页问题列表功能

+++ src/main/java/life/majiang/community/model/Question.java
@@ -4,14 +4,41 @@ import lombok.Data;
 
 @Data
 public class Question {
-    private Long id;
+    private Integer id;
+    /**
+     * 标题时间
+     */
     private String title;
+    /**
+     * 创建时间
+     */
     private Long gmtCreate;
+    /**
+     * 修改时间
+     */
     private Long gmtModified;
-    private Long creator;
+    /**
+     * 作为user表中id的外键
+     */
+    private Integer creator;
+    /**
+     * 评论数
+     */
     private Integer commentCount;
+    /**
+     * 浏览数
+     */
     private Integer viewCount;
+    /**
+     * 收藏数
+     */
     private Integer likeCount;
+    /**
+     * 标签时间
+     */
     private String tag;
+    /**
+     * 描述
+     */
     private String description;
 }
